requirements.txt
Django>=4.0,<4.1
wagtail>=3.0,<3.1
pytest-django~=4.8.0
pytest-playwright~=0.4.3
